Adding a basic parser for the proj command.
We start with 3 attribute "fields": name, git, path

The idea is we use name as an identifier for a project, git to perhaps
clone the project if it doesn't exist, and path as a way to cd and
open windows in the terminal at the expected path.

We parse everything ourselves and do not (yet) do any Tcl evaluation
within the proj proc.